test: fix Pagination test after component update
