+++ packages/docs/pages/Table/BasicTablePage.tsx
@@ -4,7 +4,31 @@ import React from 'react';
 import { CodePreview } from '../../components';
 import { TableActionsPropTable, TableBodyPropTable, TableCellPropTable, TablePropTable } from '../../PropTables';
 
-import { data } from './data';
+interface Data {
+  sku: string;
+  name: string;
+  price: number;
+}
+
+function createData(sku, name, price): Data {
+  return { sku, name, price };
+}
+
+const data: Data[] = [
+  createData('SM13', '[Sample] Smith Journal 13', 25),
+  createData('DPB', '[Sample] Dustpan & Brush', 34.95),
+  createData('OFSUC', '[Sample] Utility Caddy', 45.95),
+  createData('CLC', '[Sample] Canvas Laundry Cart', 200),
+  createData('CGLD', '[Sample] Laundry Detergent', 29.95),
+  createData('TWB', '[Sample] Tiered Wire Basket', 119.95),
+  createData('OCG', '[Sample] Oak Cheese Grater', 34.95),
+  createData('SLLPJ', '[Sample] 1 L Le Parfait Jar', 7),
+  createData('CC3C', '[Sample] Chemex Coffeemaker 3 cup', 49.5),
+  createData('ABS', '[Sample] Able Brewing System', 225),
+  createData('OTS', '[Sample] Orbit Terrarium - Small', 89),
+  createData('OTL', '[Sample] Orbit Terrarium - Large', 109),
+  createData('SLCTBS', '[Sample] Fog Linen Chambray Towel - Beige Stripe with some fondu of some sort', 49),
+];
 
 const columns: Array<{
   id: string;
@@ -61,7 +85,7 @@ export default () => {
             <>
               <Table selectable stickyHeader>
                 <Table.Actions alignItems="center" justifyContent="stretch">
-                  <Flex.Item flexGrow={2}>Test</Flex.Item>
+                  <Flex.Item flexGrow={2}>Products</Flex.Item>
                   <Flex.Item>
                     <Pagination
                       currentPage={page}
